third_party/blink/web_tests/svg/parser/whitespace-length-invalid-2-expected.txt

CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "Epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "E-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " Epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " E-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   Epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   E-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t emm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t emm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t emm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t emm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t Epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t eex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t emm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t ept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t epc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t E-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fein".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fecm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\femm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fept".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEin ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEcm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEmm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feem ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepx ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fein ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fecm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\femm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fept ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEin   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEcm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEmm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feem   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepx   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fein   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fecm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\femm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fept   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEin\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEcm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEmm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feem\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepx\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fein\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fecm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\femm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fept\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEin\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEcm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEmm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fEpc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feem\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\feex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepx\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fein\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fecm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\femm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fept\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fepc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fE-%\f".
This is a testharness.js-based test.
Harness: the test ran to completion.
